requirements.txt
Flask
Python-dotenv
Requests
Flask-oauthlib
Six 
authlib 
python-oauth2 
auth0-python 
jsonify